Update domain and other improvements.

+++ cam-service/src/main/java/it/eng/cam/rest/CAMRestImpl.java
@@ -388,6 +388,11 @@ public class CAMRestImpl {
     }
 
     public static void deleteOrionConfig(RepositoryDAO dao, String orionConfigId) {
+        List<IndividualItem> individualsByOrionConfig = dao.getIndividualsByOrionConfig(orionConfigId);
+        if (individualsByOrionConfig.size() > 0) {
+            throw new IllegalStateException("Orion Configuration cannot be deleted as it is referenced somewhere else");
+        }
+        dao = releaseRepo(dao);
         dao.deleteOrionConfig(orionConfigId);
     }
